# src/main/resources/static/openapi/inventory-api-v1.yaml
openapi: 3.0.3
info:
  title: PMIS  — Inventory API
  description: |
    REST API for Inventory Management (Safaricom Healthcare PMIS). Operational deployment
    aligns with the API gateway prefix **/api/v1**; `servers[0].url` should be appended
    before each path (for example `/inventory/products`).
  version: 1.1.0
  contact:
    name: PMIS Engineering
servers:
  - url: https://api.pmis.safaricom.com/api/v1
    description: Production (gateway)
  - url: https://api-staging.pmis.safaricom.com/api/v1
    description: Staging (gateway)
  - url: http://localhost:{port}/api/v1
    description: Local inventory service
    variables:
      port:
        default: "8080"
...
security:
  - BearerAuth: []
paths:
...
  /inventory/products/{product_id}/batches:
    get:
      tags: [Batches]
      operationId: listBatches
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BatchStatus'
        - name: sort
          in: query
          schema:
            type: string
            enum: [fefo, lifo, created_at]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchListResponse'
    post:
      tags: [Batches]
      operationId: addBatch
      summary: Add batch (JSON) or bulk CSV (multipart)
      description: |
        **application/json** → response body is a single `Batch`.

        **multipart/form-data** (`file` required) → response body is **`BatchBulkUploadResult`**
        (manifest of created rows + per-row failures).
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: JSON returns **Batch**; CSV returns **BatchBulkUploadResult**
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Batch'
                  - $ref: '#/components/schemas/BatchBulkUploadResult'
  /inventory/products/{product_id}/batches/{batch_id}:
    get:
      tags: [Batches]
      operationId: getBatch
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/BatchIdParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
    delete:
      tags: [Batches]
      operationId: removeBatch
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/BatchIdParam'
      responses:
        '204':
          description: Removed
  /inventory/products/{product_id}/adjustments:
    get:
      tags: [Stock Adjustments]
      operationId: listAdjustments
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdjustmentListResponse'
    post:
      tags: [Stock Adjustments]
      operationId: adjustStock
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockAdjustment'
...
  /inventory/purchase-orders:
    get:
      tags: [Purchase Orders]
      operationId: listPurchaseOrders
      summary: List purchase orders for the tenant
      description: |
        Response bodies use **camelCase** JSON properties (matches runtime Java records).
        Send **`X-Tenant-Id`** when the JWT does not carry tenant context (see implementation map).
      responses:
        '200':
          description: Purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PurchaseOrderResponse'
    post:
      tags: [Purchase Orders]
      operationId: createPurchaseOrder
      summary: Create a purchase order in ERPNext
      description: Request body uses **camelCase** JSON properties.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductIdParam:
      name: product_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BatchIdParam:
      name: batch_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    BatchStatus:
      type: string
      enum:
        [available, active, expired, recalled, quarantine, unavailable]
    Pagination:
      type: object
      properties:
        page: {type: integer}
        limit: {type: integer}
        total: {type: integer}
        total_pages: {type: integer}
...
    Batch:
      type: object
      additionalProperties: true
    CreateBatchRequest:
      type: object
      additionalProperties: true
    BatchListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
        pagination:
          $ref: '#/components/schemas/Pagination'
    BatchBulkUploadResult:
      type: object
      properties:
        total_rows: {type: integer}
        success_count: {type: integer}
        error_count: {type: integer}
        errors:
          type: array
          items:
            type: object
        created_batches:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
    StockAdjustmentRequest:
      type: object
      additionalProperties: true
    StockAdjustment:
      type: object
      additionalProperties: true
    AdjustmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/StockAdjustment'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PurchaseOrderLineItem:
      type: object
      required: [itemCode, qty]
      properties:
        itemCode:
          type: string
          description: ERPNext Item code
        qty:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        rate:
          type: number
          format: double
          description: Unit rate (optional)
    CreatePurchaseOrderRequest:
      type: object
      required: [supplier, items]
      properties:
        supplier:
          type: string
          description: ERPNext Supplier document name
        transactionDate:
          type: string
          description: Transaction date YYYY-MM-DD
        scheduleDate:
          type: string
          description: Expected delivery date YYYY-MM-DD
        notes:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PurchaseOrderLineItem'
    PurchaseOrderResponse:
      type: object
      properties:
        id: { type: string, description: ERPNext Purchase Order name }
        poNumber: { type: string }
        supplier: { type: string }
        orderDate: { type: string }
        expectedDate: { type: string }
        receivedDate: { type: string, nullable: true }
        status: { type: string }
        items: { type: integer, description: Line item count }
        totalAmount: { type: number, format: double }
...